Added possibility to initialise a gaussian sphere of fluid.

+++ initial.c
@@ -365,13 +365,13 @@ void fluid_sphere(hydro_fields f, hydro_params p){
 
   int x, y, z;
   
-  int direct_x, direct_y, direct_z;
-  int wrap_x, wrap_y, wrap_z;
-  int delta_x, delta_y, delta_z;
+  float direct_x, direct_y, direct_z;
+  float wrap_x, wrap_y, wrap_z;
+  float delta_x, delta_y, delta_z;
   
-  int x0 = 0;
-  int y0 = 0;
-  int z0 = 0;
+  float x0 = 0;
+  float y0 = 0;
+  float z0 = 0;
   
   for(x = 1; x <= p.slicex; x++) {
     for(y = 1; y <= p.slicey; y++) {
